% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node_geoloc.R
\name{add_node_geoloc}
\alias{add_node_geoloc}
\title{Add coordinates on nodes in a SIC project XML file}
\usage{
add_node_geoloc(x, y, cfg)
}
\arguments{
\item{x, y}{\link{numeric}, coordinates of the nodes in the SIC project}

\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}
}
\value{
Nothing. Use for side effect on the XML project file.
}
\description{
Add coordinates on nodes in a SIC project XML file
}
