% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_tmp_project.R
\name{cfg_tmp_project}
\alias{cfg_tmp_project}
\title{Set a configuration with a temporary project directory}
\usage{
cfg_tmp_project(
  xml_path = system.file("sic_project_test1.xml", package = "rsic2"),
  cfg = loadConfig(xml_path = xml_path)
)
}
\arguments{
\item{xml_path}{\link{character}, the path of the XML SIC project file}

\item{cfg}{\link{config}, the configuration to modify}
}
\value{
The updated configuration with the temporary project directory
}
\description{
Copy the XML SIC project file in a temporary folder and modify the configuration
according the new location of the XML SIC project file.
}
\examples{
# Copy the example project provided in the package into a temporary folder
cfg <- cfg_tmp_project()
cfg$project$xml_path

}
