% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_xml_section.R
\name{copy_xml_section}
\alias{copy_xml_section}
\title{Copy an XML section from one reach to another}
\usage{
copy_xml_section(
  source_reach,
  source_section,
  target_reach,
  target_section,
  xp = xml2::read_xml(cfg$project$path),
  write_xml = TRUE,
  cfg
)
}
\arguments{
\item{source_reach}{\link{integer} the reach number of the source section}

\item{source_section}{\link{integer} the section number of the source reach}

\item{target_reach}{\link{integer} the reach number of the target section}

\item{target_section}{\link{integer} the section number of the target reach}

\item{xp}{A \code{xml_document} of the XML project file. Default is the one in the configuration}

\item{write_xml}{\link{logical} if \code{TRUE}, the XML file is saved after modification (Needs that \code{cfg} is provided)}

\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}
}
\value{
Nothing. Use for side effect on the XML project file.
}
\description{
Copy an XML section from one reach to another
}
