% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_uniform_reach_txt.R
\name{create_uniform_reach_txt}
\alias{create_uniform_reach_txt}
\title{Generate text export geometry for an uniform reach}
\usage{
create_uniform_reach_txt(
  abscissas,
  upstream_bed_elevation,
  slope,
  section_type,
  profile,
  section_names = paste0("Section x=", abscissas)
)
}
\arguments{
\item{abscissas}{\link{numeric} vector of section abscissas}

\item{upstream_bed_elevation}{\link{numeric}, upstream bed elevation (m)}

\item{slope}{\link{numeric}, bed slope of the reach (m/m)}

\item{section_type}{1-length \link{character}, type of the section: "T" for Trapezoidal, "A" for Abscissa/Elevation, "L" for Width/Elevation}

\item{profile}{\link{list} or \link{matrix}, profile of the section (See details)}

\item{names}{\link{character} vector of section names}
}
\value{
A \link{list} from which each item is a section exported by \link{create_section_txt}.
Names of the list are the abscissas with trailing zeros for character sorting.
}
\description{
Generate text export geometry for an uniform reach
}
