% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_to_reach.R
\name{dem_to_reach_txt}
\alias{dem_to_reach_txt}
\alias{dem_to_reach}
\title{Create SIC geometry sections from a DEM}
\usage{
dem_to_reach_txt(
  dem,
  node_coords,
  space_step,
  section_width,
  nb_points = 50,
  start = 0,
  major_bed = FALSE
)

dem_to_reach(dem, node_coords, section_centers, section_width, nb_points = 50)
}
\arguments{
\item{dem}{\link[terra:SpatRaster-class]{terra::SpatRaster} object with altitude data (m).}

\item{node_coords}{\link{matrix} with coordinates of starting and ending points of the line to browse}

\item{space_step}{1-length \link{numeric}, distance between each section (m)}

\item{section_width}{1-length \link{numeric}, width of the sections to create}

\item{nb_points}{1-length \link{numeric}, number of points to describe cross-section geometries}

\item{start}{1-length \link{numeric}, starting value for the chainage (i.e. section abscissa) along the reach}

\item{major_bed}{\link{logical}, \code{TRUE} for major bed, \code{FALSE} for minor-medium bed}

\item{section_centers}{See return value of \link{get_section_centers}}
}
\value{
A \emph{ReachTxt} object which is a \link{list} of \emph{SectionTxt} objects (see \link{create_section_txt}).
}
\description{
The coordinate system of \code{dem} should be a metric orthonormal coordinate system.
}
\examples{
## Inputs preparation
data("floodam_ead_dem")
dem <- terra::rast(floodam_ead_dem)
node_coords <- matrix(c(102550, 102550, 110000, 108000), ncol = 2)
space_step = 100
section_width = 5000

## Get section positions
reach_length <- as.numeric(dist(node_coords[1:2,]))
nb_sections <- ceiling(reach_length / space_step) + 1
section_centers <- get_section_centers(node_coords, nb_sections)

## Create a section profile and plot it!
profile <- dem_to_section(dem, node_coords, section_centers[5,], section_width)
plot(profile)

## Generate cross section profiles for a reach
reach <- dem_to_reach(dem, node_coords, section_centers, section_width)
plot(reach[[5]])

## Generate section profiles in SIC import text format
reach_txt <- dem_to_reach_txt(dem, node_coords, space_step, section_width, major_bed = TRUE)
reach_txt[[5]]

}
