% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_reach.R
\name{extract_reach}
\alias{extract_reach}
\title{Select portion of a reach between two chainages}
\usage{
extract_reach(reach, x_limits)
}
\arguments{
\item{reach}{A ReachTxt object}

\item{x_limits}{2-length \link{numeric}, min and max chainage}
}
\value{
A ReachTxt object containing the selected sections.
}
\description{
Select portion of a reach between two chainages
}
