% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.pathwin.R
\name{file.pathwin}
\alias{file.pathwin}
\title{file.path for Windows}
\usage{
file.pathwin(...)
}
\arguments{
\item{...}{Arguments sent to \link{file.path}}
}
\description{
Same behavior as \link{file.path} but remove duplicated slashes and use backslashes as the path separator.
}
