% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_to_reach.R
\name{get_section_centers}
\alias{get_section_centers}
\title{Get coordinates of section centers}
\usage{
get_section_centers(node_coords, nb_sections)
}
\arguments{
\item{node_coords}{2-rows \link{matrix}, with coordinates of segment boundaries}

\item{nb_sections}{1-length \link{numeric}, number of points to extract}
}
\value{
A \link{matrix} with the coordinates of the points along the segment
}
\description{
From the coordinates of a segment, this function returns the coordinates
of equidistant points along the segment given a number of points
}
\examples{
node_coords <- matrix(c(102550, 102550, 110000, 108000), ncol = 2)
reach_length <- as.numeric(dist(node_coords[1:2,]))
space_step <- 100
nb_sections <- ceiling(reach_length / space_step) + 1
get_section_centers(node_coords, nb_sections)
}
