% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_geoloc.R
\name{get_xml_geoloc}
\alias{get_xml_geoloc}
\title{Retrieve the geolocation coordinates of a node from an XML node}
\usage{
get_xml_geoloc(xml_node)
}
\arguments{
\item{xml_node}{A XML node of a SIC node or section}
}
\value{
A numeric vector with the geolocation coordinates of the node
}
\description{
Retrieve the geolocation coordinates of a node from an XML node
}
