% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_num_by_name.R
\name{get_xml_node_num_by_name}
\alias{get_xml_node_num_by_name}
\alias{get_xml_reach_num_by_name}
\title{Retrieve node or reach number from XML project file}
\usage{
get_xml_node_num_by_name(name, xp = xml2::read_xml(cfg$project$path), cfg)

get_xml_reach_num_by_name(name, xp = xml2::read_xml(cfg$project$path), cfg)
}
\arguments{
\item{name}{\link{character} name of the node or reach}

\item{xp}{\link[xml2:xml_document]{xml2::xml_document} the XML project file. Default is the one in the configuration}

\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}
}
\value{
\link{integer} the number of the node or reach in the XML project file
}
\description{
Retrieve node or reach number from XML project file
}
