% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_num_by_name.R
\name{get_xml_reach_num_by_node_num}
\alias{get_xml_reach_num_by_node_num}
\title{Get reach number from node number}
\usage{
get_xml_reach_num_by_node_num(
  node_num,
  upstream = TRUE,
  xp = xml2::read_xml(cfg$project$path),
  cfg
)
}
\arguments{
\item{node_num}{\link{numeric} node number in the SIC model}

\item{upstream}{\link{logical} if TRUE, get the reach number from the upstream node, otherwise downstream}

\item{xp}{A \code{xml_document} of the XML project file. Default is the one in the configuration}

\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}
}
\value{
\link{integer} the reach number in the XML project file
}
\description{
Get reach number from node number
}
\details{
In case of several XML nodes are matching, the first one is returned.
Providing \code{cfg} is not necessary if \code{xp} is provided.
}
\examples{
xp <- xml2::read_xml(system.file("sic_project_test1.xml", package = "rsic2"))
get_xml_reach_num_by_node_num(1, upstream = TRUE, xp)
get_xml_reach_num_by_node_num(2, upstream = FALSE, xp)
}
