% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{loadConfig}
\alias{loadConfig}
\title{Read default configuration of the package and complete it with eventual user config}
\usage{
loadConfig(
  sic_path = NULL,
  xml_path = NULL,
  userFile = "config.yml",
  pathDefaultCfg = system.file("config.yml", package = "rsic2")
)
}
\arguments{
\item{userFile}{location of the user config YML file}

\item{pathDefaultCfg}{The location of the default configuration (located in "inst/config.yml" of the package by default)}
}
\value{
A configuration as it is returned by \link[config:get]{config::get}.

Configuration of RSIC2 as the following structure:
\itemize{
\item sic
\itemize{
\item path: Path of local SIC installation (should be defined by \code{sic_path} parameter or in the user config file)
\item edisic: sub-path to EdiSIC program
\item talweg: sub-path to TALWEG program
\item fluvia: sub-path to FLUVIA program
\item sirene: sub-path to SIRENE program
\item export: sub-path to SicExport program
\item fortran:
\itemize{
\item prms:
\itemize{
\item INTERF: default \code{INTERF} parameter injected in command line arguments of TALWEG, FLUVIA, SIRENE
}
}
}
\item project
\itemize{
\item path: Path to the XML project file (should be defined by \code{xml_path} parameter or in the user config file)
}
}
}
\description{
Read default configuration of the package and complete it with eventual user config
}
\examples{
library(sic2)
sic_path <- tempdir(check = TRUE)
xml_path <- R.utils::tmpfile()
cfg <- loadConfig(sic_path, xml_path)
str(cfg)
}
