% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_inputs.R
\name{merge.SicInput}
\alias{merge.SicInput}
\title{Merge SicInput objects into a list}
\usage{
\method{merge}{SicInput}(x, y = NULL, ...)
}
\arguments{
\item{x}{a \link{SicInput} object to merge}

\item{y}{a \link{SicInput} object to merge}

\item{...}{other \link{SicInput} objects to merge}
}
\value{
A \code{SicInputs} object which is a list of \link{SicInput}
}
\description{
Merge SicInput objects into a list
}
\details{
The list of parameter is compliant with the \code{merge} S3 method
available in R, so this method can either be called by typing
\code{merge} or \code{merge.SicInput}.
}
\examples{
# How to impose 5 m3/s in the upstream offtake of the model?
# Define location of the boundary condition to set
locations <- SicLocation(list(Nd = 1, Pr = 1, car = "Q"))
# Define its value
sicInputUpstream <- SicInput(5, locations = locations)

# Opening a gate at 0.5 m
sicInputGate <- SicInput(
  0.5,
  locations = SicLocations(list(Bf = 3, Sn = 2, Ouv = 1, Car = "Ouverture"))
)

# Merging all inputs
sicInputs <- merge(sicInputUpstream, sicInputGate)

}
