% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_boundary_ZQ.R
\name{set_boundary_ZQ}
\alias{set_boundary_ZQ}
\title{Set stage-discharge curve at an offtake of the model}
\usage{
set_boundary_ZQ(cfg, scenario, nd, pr = 1, mZQ)
}
\arguments{
\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}

\item{scenario}{\link{numeric}, the scenario to use}

\item{nd}{\link{numeric}, node number in SIC model}

\item{pr}{\link{numeric}, offtake number in the node}

\item{mZQ}{\link{matrix}, stage-discharge relation with first column the water elevation in meters
and the second column the discharge in cubic meters per seconds}
}
\value{
This function is only used for side effect on the XML file of the SIC project.
}
\description{
Update the existing stage-discharge boundary condition.
}
\examples{
\dontrun{
cfg <- cfg_tmp_project()
z <- seq(5, 20, 1)
Q <- (z-5)^1.5
set_boundary_ZQ(cfg, scenario = 1, nd = 3, mZQ = matrix(c(z, Q), ncol = 2))
}
}
