% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_initial_conditions.R
\name{set_initial_conditions}
\alias{set_initial_conditions}
\title{Import initial conditions}
\usage{
set_initial_conditions(iniParams, cfg = loadConfig())
}
\arguments{
\item{iniParams}{\link{numeric}, arguments passed to Edisic for importing initial conditions. See details}

\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}
}
\value{
Error code returned by \link{shell}.
}
\description{
Import initial conditions from a scenario/variant result calculation to a new scenario/variant.
}
\details{
\code{iniParams} \link{numeric} vector of length 5. Each number correspond to:
\enumerate{
\item number of the scenario from which import initial conditions
\item number of the variant from which import initial conditions (Use "0" for no variant)
\item Time in seconds of the initial conditions to import (Use "0" for a single permanent simulation with no time steps)
\item number of the scenario where the initial conditions are exported to
\item number of the variant where the initial conditions are exported to  (Use "0" for no variant)
}
}
\examples{
\dontrun{
# Set up the configuration model
cfg <- cfg_tmp_project()

# Generate the mesh of the model
sic_run_mesh(cfg)

# Run steady simulation for the scenario #1
sic_run_steady(cfg, scenario = 1)

# Import initial condition from scenario 1
# to scenario 1, variant 1 for unsteady flow simulation
set_initial_conditions(c(1, 0, 0, 1, 1), cfg = cfg)

# Run unsteady flow simulation
sic_run_unsteady(cfg, scenario = 1, variant = 1)

# Or initiate and run the same unsteady flow simulation in one call
sic_run_unsteady(cfg, iniParams = c(1, 0, 0, 1, 1))
}
}
