% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_project_params.R
\name{set_project_params}
\alias{set_project_params}
\title{Set project params in the SIC project XML file}
\usage{
set_project_params(params, cfg)
}
\arguments{
\item{params}{\link{list} of parameters (See details)}

\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}
}
\value{
None. Use for side effect on the XML project file.
}
\description{
Set project params in the SIC project XML file
}
\details{
The parameter \code{params} is a list of parameters to set in the SIC project XML file.
The list should be named with the parameter name.
Possible values of parameters are:
\itemize{
\item \code{HomothetieX}
\item \code{HomothetieY}
\item \code{DX}
\item \code{ImportEdital}
\item \code{GeoUnit}
\item \code{GeoRefAngle}
}
}
