% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_run_fortran.R
\name{sic_run_fortran}
\alias{sic_run_fortran}
\title{Run Talweg, Fluvia or Sirene}
\usage{
sic_run_fortran(prog, params = list(), cfg = loadConfig())
}
\arguments{
\item{prog}{\link{character}, the program to run. Should be one of "talweg"}

\item{params}{\link{list} or \link{character}, see details}

\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}
}
\value{

}
\description{
Run Talweg, Fluvia or Sirene
}
\details{
If argument \code{params} is a \link{list}, arguments are injected in the command line by taking the items of the list with the conversion
\verb{[key]=[value]}. If argument \code{params} is a \link{character}
}
\examples{
\dontrun{
# Run steady simulation for the scenario #1
params <- list(SCE=1)
sic_run_fortran("fluvia", params)
}
}
