% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_inputs.R
\name{sic_write_par}
\alias{sic_write_par}
\title{Write a PAR file for SIC simulation}
\usage{
sic_write_par(cfg, scenario, sicInputs)
}
\arguments{
\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}

\item{scenario}{\link{numeric}, the scenario to use}

\item{sicInputs}{A \link{SicInput} object or a list of \link{SicInput} objects create by \link{merge.SicInput}
used to create a PAR file injectig inputs for the simulation}
}
\value{
Nothing.
}
\description{
Write inputs  in a PAR file respectively to \url{https://sic.g-eau.fr/Format-of-the-par-file}.
}
\examples{
\dontrun{
# Setting the model configuration
cfg <- cfg_tmp_project()

# How to impose an hydrograph at the upstream offtake of the model?
# Define the location of the upstream boundary condition to set
locations <- SicLocations(list(Nd = 1, Pr = 1, car = "Q"))
# Define its time series
sicInputs <- SicInput(data.frame(t = c(0, 3600, 7200), # time in seconds
                                v = c(5, 20, 5)),     # flows
                      locations = locations)
# Write the parameters in the PAR file
sic_write_par(cfg, 1, sicInputs)
}
}
