% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_inputs.R
\name{SicInput}
\alias{SicInput}
\alias{SicInput.numeric}
\alias{SicInput.POSIXt}
\alias{SicInput.data.frame}
\alias{SicInput.matrix}
\title{Create a SIC input for a PAR file}
\usage{
SicInput(x, ...)

\method{SicInput}{numeric}(x, values = NULL, locations, interpolated = TRUE, ...)

\method{SicInput}{POSIXt}(x, values, start = NULL, ...)

\method{SicInput}{data.frame}(x, ...)

\method{SicInput}{matrix}(x, ...)
}
\arguments{
\item{x}{either a fixed input, a \link{numeric} vector of time in seconds,
a \link{POSIXt} vector of time, a \link{matrix} or a \link{data.frame} with 2 columns (time and value)}

\item{...}{used for S3 method compatibility}

\item{values}{a \link{numeric} vector used if \code{x} is a vector of \link{numeric} or \link{POSIXt}}

\item{locations}{input locations created with \link{SicLocation} or \link{SicLocations}}

\item{interpolated}{Interpolation mode \code{TRUE} for "ramp" mode and \code{FALSE} for "step" mode}

\item{start}{a \link{POSIXt} indicating the start time to use as time zero in the simulation}
}
\value{
A \emph{SicInput} object which is a list with the following items:
\itemize{
\item \code{locations}: a \link{SicLocations} object
\item \code{data}: \link{numeric}, the fixed value, or \link{data.frame}, the time series to apply to the locations
\item \code{interpolated}: \link{logical}, interpolation mode
}
}
\description{
Create a SIC input for a PAR file
}
\examples{
# How to impose 5 m3/s in the upstream offtake of the model?
# Define location of the boundary condition to set
locations <- SicLocations(list(Nd = 1, Pr = 1, car = "Q"))
# Define its value
sicInputs <- SicInput(5, locations = locations)

# How to impose an hydrograph at the upstream offtake of the model?
sicInputs <- SicInput(data.frame(t = c(0, 3600, 7200), # time in seconds
                                v = c(5, 20, 5)),     # flows
                      locations = locations)
}
\seealso{
Other SicInput: 
\code{\link{SicLocation}()},
\code{\link{SicLocations}()}
}
\concept{SicInput}
