% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_inputs.R
\name{SicLocations}
\alias{SicLocations}
\title{Set locations of a SIC model input}
\usage{
SicLocations(...)
}
\arguments{
\item{...}{One or several \link{list} describing a location (See \link{SicLocation})}
}
\value{
a \emph{SicLocations} object which is a \link{list} of \link{SicLocation}.
}
\description{
Do the same as \link{SicLocation} for eventually several locations
}
\examples{
# Applying the same flow to offtakes located in nodes number 1 to 10
locations <- lapply(seq(10), function(i) { list(Nd = i, Pr = 1, Car = "Q")})
locations <- do.call(SicLocations, locations)
sicInputOfftakes <- SicInput(-0.5, locations = locations)
}
\seealso{
Other SicInput: 
\code{\link{SicInput}()},
\code{\link{SicLocation}()}
}
\concept{SicInput}
