% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_reach.R
\name{split_reach}
\alias{split_reach}
\title{Split a reach into a list of reaches}
\usage{
split_reach(reach, x_limits)
}
\arguments{
\item{reach}{A ReachTxt object}

\item{x_limits}{Chainage (i.e. abscissas along the reach) of splitting points}
}
\value{
A \link{list} of ReachTxt objects
}
\description{
Split a reach into a list of reaches
}
\examples{
# Create a 10km long trapezoidal uniform reach
profT <- list(
  B = 2,
  S = (6 - 2) / 2 / 2,
  ZF = 100,
  ZB = 100 + 2
)
min_reach <- create_uniform_reach_txt(abscissas = seq(0, 10000, 100),
                                      upstream_bed_elevation = 10 + 2000 * 0.002,
                                      slope = 0.002,
                                      section_type = "T",
                                      profile = profT)
# Split into reaches of 2000 m
reaches <- split_reach(min_reach, seq(0, 10000, 2000))

}
