% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_result.R
\name{tidy_result}
\alias{tidy_result}
\title{Tidy a result simulation}
\usage{
tidy_result(res)
}
\arguments{
\item{res}{a \emph{SicResult} \link{matrix} provided by \link{get_result}}
}
\value{
A \link{data.frame} with one line per saved simulation result time step :
\itemize{
\item one column per object type of the result (example: "bf" and "sn" for a section or "nd" and "pr" for an offtake)
\item one column "var" for the definition of the result variable
\item one column "t" for the simulation time of the result variable
\item one columne "value" for the value of the result variable
}
}
\description{
Tidy a result simulation
}
