% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_nodes_names.R
\name{update_nodes_names}
\alias{update_nodes_names}
\title{Update nodes names in a SIC model XML file}
\usage{
update_nodes_names(new_names, cfg)
}
\arguments{
\item{new_names}{\link{character}, new names for the nodes in the SIC model}

\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}
}
\value{
Nothing. Use for side effect on the XML project file.
}
\description{
Update nodes names in a SIC model XML file
}
