% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_import_reaches.R
\name{update_portion_abscissas}
\alias{update_portion_abscissas}
\title{Update abscissas of Strickler portions in a SIC model}
\usage{
update_portion_abscissas(cfg, stricklers = cfg$project$stricklers)
}
\arguments{
\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}

\item{stricklers}{2-length \link{numeric}, Strickler coefficient to apply for minor bed and medium bed respectively}
}
\value{
Use for side effect on the XML project file.
}
\description{
As \link{sic_import_reaches} redefines abscissas in the reaches, and as the Stricklers
are defined with abscissas, geometry importation can lead to an inconsistent model.

So the aim of this function is to reset Strickler definitions of the whole model
to a default value with one Strickler portion by reach with correct abscissas.
}
\examples{
\dontrun{
cfg <- cfg_tmp_project()
update_portion_abscissas(cfg, KMin = 50)

# Display first Strickler definitions encountered in the XML file
x <- read_xml(cfg$project$path)
xml_find_first(x, "//Stricklers")
}
}
\concept{sic_import_reaches}
