% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sic_params.R
\name{convert_sic_params}
\alias{convert_sic_params}
\title{Convert a list of parameters into command line arguments}
\usage{
convert_sic_params(params, cfg = loadConfig())
}
\arguments{
\item{params}{a \link{list} of \link{character} with each item is identified by}

\item{cfg}{}
}
\value{
A \link{character} with each parameter is converted to "\link{key param}=\link{value param}" and each parameter separated by a space character.
}
\description{
Convert a list of parameters into command line arguments
}
