% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_section_txt.R
\name{create_section_txt}
\alias{create_section_txt}
\title{Export a section in importation SIC format}
\usage{
create_section_txt(
  section_name,
  abscissa,
  section_type,
  profile,
  distance_majeur = FALSE
)
}
\arguments{
\item{section_name}{\link{character}, name of the section}

\item{abscissa}{\link{numeric}, abscissa of the section in the reach}

\item{section_type}{1-length \link{character}, type of the section: "T" for Trapezoidal, "A" for Abscissa/Elevation, "L" for Width/Elevation}

\item{profile}{\link{list} or \link{matrix}, profile of the section (See details)}

\item{distance_majeur}{\link{logical} or \link{numeric}, \code{FALSE} for a minor bed section}
}
\value{
\link{character} or \link{list} depending on the \code{as_list} parameter.
}
\description{
Export a section in importation SIC format
}
\examples{
# Trapezoidal section
export_section_txt("Trapeze", 0, "T", list(B = 2, S = 1.5, ZF = 100, ZB = 102))
}
