% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodam_ead_dem.R
\docType{data}
\name{floodam_ead_dem}
\alias{floodam_ead_dem}
\title{Default DEM from the floodam-ead package}
\format{
A \code{PackedSpatRaster} object, use 'terra::rast()' to unpack it.
}
\source{
https://gitlab.irstea.fr/flood-appraisal/floodam/floodam-ead
}
\usage{
floodam_ead_dem
}
\description{
This is the default DEM generated by the function \code{generate_DEM} located in
the data-raw folder of the package source code, with default arguments.
}
\examples{
data("floodam_ead_dem", package = "rsic2")
floodam_ead_dem
terra::rast(floodam_ead_dem)
}
\keyword{datasets}
