% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R
\name{get_result}
\alias{get_result}
\title{Get resultat}
\usage{
get_result(
  cfg,
  scenario,
  variant = 0,
  filters = c(""),
  m = read_bin_result_matrix(cfg, scenario, variant)
)
}
\arguments{
\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}

\item{scenario}{\link{numeric}, the scenario to read}

\item{variant}{\link{numeric}, the variant to read}

\item{m}{}
}
\value{

}
\description{
Get resultat
}
\examples{
cfg <- cfg_tmp_project()
sic_run_fortran("fluvia", list(SCE = 1), cfg = cfg)
get_result(cfg, 1, filters = c("bf==4", "var=='Z'"))
}
