% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R
\name{get_result}
\alias{get_result}
\title{Get a selection of variables from a simulation result}
\usage{
get_result(
  cfg,
  scenario,
  variant = 0,
  filters = c(""),
  m = read_bin_result_matrix(cfg, scenario, variant)
)
}
\arguments{
\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}

\item{scenario}{\link{numeric}, the scenario to read}

\item{variant}{\link{numeric}, the variant to read}

\item{filters}{\link{character} conditions to select columns in result table, see details}

\item{m}{\link{matrix} of results produced by \link{read_bin_result_matrix}}
}
\value{
\link{matrix} of results with columns selected by \code{filters}.
}
\description{
Get a selection of variables from a simulation result
}
\examples{
cfg <- cfg_tmp_project()
sic_run_fortran("fluvia", list(SCE = 1), cfg = cfg)
get_result(cfg, 1, filters = c("bf==4", "var=='Z'"))
}
