% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R
\name{get_result}
\alias{get_result}
\title{Get a selection of variables from a simulation result}
\usage{
get_result(
  cfg,
  scenario,
  variant = 0,
  filters = c(""),
  fun_format = NULL,
  m = read_bin_result_matrix(cfg, scenario, variant)
)
}
\arguments{
\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}

\item{scenario}{\link{numeric}, the scenario to use}

\item{variant}{\link{numeric}, the variant to use (0 by default means no variant)}

\item{filters}{\link{character} conditions to select columns in result table, see details}

\item{fun_format}{\link{function} to format the result (See \link{tidy_result})}

\item{m}{\link{matrix} of results produced by \link{read_bin_result_matrix}}
}
\value{
If \code{format =NULL}, it's a \link{matrix} of results with a first column "t" with the simulation time
in seconds followed by columns selected by \code{filters}.

Column names are a concatenation of nested SIC model elements separated by
the character "|" and numbered after the character ":". The variable is
represented by the item "var".
For example, water elevation in the first section of the first reach is:
"bf:1|sn:1|var:Z".

If \code{format = tidy_result} or \code{format = compact_tidy_result}, see the documentation of \link{tidy_result}.
}
\description{
Get a selection of variables from a simulation result
}
\examples{
\dontrun{
cfg <- cfg_tmp_project()
sic_run_steady(cfg, scenario = 1)
get_result(cfg, 1, filters = c("bf==4", "var=='Z'"))
}
}
