% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_initial_conditions.R
\name{set_initial_conditions}
\alias{set_initial_conditions}
\title{Import initial conditions}
\usage{
set_initial_conditions(params, cfg = loadConfig())
}
\arguments{
\item{params}{\link{numeric}, arguments passed to Edisic for importing initial conditions. See details}

\item{cfg}{}
}
\value{

}
\description{
Import initial conditions from a scenario/variant result calculation to a new scenario/variant.
}
\details{
\code{params} \link{numeric} vector of length 5. Each number correspond to:
\enumerate{
\item number of the scenario from which import initial conditions
\item number of the variant from which import initial conditions (Use "0" for no variant)
\item Time in seconds of the initial conditions to import (Use "0" for a single permanent simulation with no time steps)
\item number of the scenario where the initial conditions are exported to
\item number of the variant where the initial conditions are exported to  (Use "0" for no variant)
}
}
\examples{
\dontrun{
# Import initial conditions from the scenario #1 without variant at time 0s
# to the variant #1 in the scenario #1
set_initial_conditions("1 0 0 1 1")
}
}
