% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_run_export.R
\name{sic_run_export}
\alias{sic_run_export}
\title{Run SicExport and read the exported file}
\usage{
sic_run_export(scenario, variant = 0, params, cfg = loadConfig())
}
\arguments{
\item{scenario}{\link{numeric}, the scenario to read}

\item{variant}{\link{numeric}, the variant to read}

\item{params}{\link{list} location parameters of the result, see details.}

\item{cfg}{a \link{config} object. Configuration to use. See \link{loadConfig} for details}
}
\value{
\link{matrix} with the read result
}
\description{
Run SicExport and read the exported file
}
\details{
\code{params} parameter is a list representing parameters available in \url{https://sic.g-eau.fr/sicexport-utilitaire-d-exportation} to set the model network location of exported results. The string parameter \verb{/x=n /yy=ii} in the command line is here represented by \code{list(xxx = nnn, yy = ii)}.
}
\examples{
\dontrun{
params <- list(SCE=1)
sic_run_fortran("fluvia", params)
# For exporting result in sections at time 0
sic_run_export(scenario = 1, params = list(t = 0))
}
}
