% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_reach.R
\name{split_reach}
\alias{split_reach}
\title{Split a reach into a list of reaches}
\usage{
split_reach(reach, x_limits)
}
\arguments{
\item{reach}{A ReachTxt object}

\item{x_limits}{Chainage (i.e. abscissas along the reach) of splitting points}
}
\value{
A \link{list} of ReachTxt objects
}
\description{
Split a reach into a list of reaches
}
